refactor :응답 response 래퍼타입들 기본타입으로 수정
