+++ src/main/java/cart/service/product/dto/ProductResponse.java
@@ -3,12 +3,12 @@ package cart.service.product.dto;
 import cart.service.product.domain.Product;
 
 public class ProductResponse {
-    private final Long id;
+    private final long id;
     private final String name;
     private final String imageUrl;
-    private final Integer price;
+    private final int price;
 
-    public ProductResponse(Long id, String name, String imageUrl, Integer price) {
+    public ProductResponse(Long id, String name, String imageUrl, int price) {
         this.id = id;
         this.name = name;
         this.imageUrl = imageUrl;
@@ -31,7 +31,7 @@ public class ProductResponse {
         return imageUrl;
     }
 
-    public Integer getPrice() {
+    public int getPrice() {
         return price;
     }
 }
